+++ tests/test_managers.py
@@ -56,6 +56,16 @@ class TestResourceManager(unittest.TestCase):
         self.assertEqual(project.identifier, 'foo')
         self.assertEqual(project.id, 1)
 
+    @mock.patch('redmine.requests.get')
+    def test_get_unicode_resource(self, mock_get):
+        mock_get.return_value = response = mock.Mock(status_code=200)
+        unicode_name = b'\xcf\x86oo'.decode('utf8')
+        response.json.return_value = {'project': {'name': unicode_name, 'identifier': unicode_name, 'id': 1}}
+        project = self.redmine.project.get(unicode_name)
+        self.assertEqual(project.name, unicode_name)
+        self.assertEqual(project.identifier, unicode_name)
+        self.assertEqual(project.id, 1)
+
     def test_get_all_resources(self):
         self.assertIsInstance(self.redmine.project.all(), ResourceSet)
 
@@ -78,6 +88,15 @@ class TestResourceManager(unittest.TestCase):
         user = self.redmine.user.create(firstname='John', lastname='Smith')
         self.assertEqual(user.firstname, 'John')
         self.assertEqual(user.lastname, 'Smith')
+
+    @mock.patch('redmine.requests.post')
+    def test_create_unicode_resource(self, mock_post):
+        mock_post.return_value = response = mock.Mock(status_code=201)
+        unicode_name = b'\xcf\x86oo'.decode('utf8')
+        response.json.return_value = {'wiki_page': {'title': unicode_name, 'project_id': 1}}
+        wiki_page = self.redmine.wiki_page.create(title=unicode_name, project_id=1)
+        self.assertEqual(wiki_page.title, unicode_name)
+        self.assertEqual(wiki_page.project_id, 1)
 
     @mock.patch('redmine.open', mock.mock_open(), create=True)
     @mock.patch('redmine.requests.post')
@@ -102,7 +121,7 @@ class TestResourceManager(unittest.TestCase):
         mock_put.return_value = mock.Mock(status_code=200, content='')
         manager = self.redmine.wiki_page
         manager.params['project_id'] = 1
-        self.assertEqual(self.redmine.wiki_page.update('Foo', project_id=1, title='Bar'), True)
+        self.assertEqual(manager.update(b'\xcf\x86oo'.decode('utf8'), title='Bar'), True)
 
     @mock.patch('redmine.open', mock.mock_open(), create=True)
     @mock.patch('redmine.requests.put')
@@ -118,7 +137,7 @@ class TestResourceManager(unittest.TestCase):
     @mock.patch('redmine.requests.delete')
     def test_delete_resource(self, mock_delete):
         mock_delete.return_value = mock.Mock(status_code=200, content='')
-        self.assertEqual(self.redmine.group.delete(1), True)
+        self.assertEqual(self.redmine.wiki_page.delete(b'\xcf\x86oo'.decode('utf8'), project_id=1), True)
 
     def test_resource_get_method_unsupported_exception(self):
         self.assertRaises(ResourceBadMethodError, lambda: self.redmine.enumeration.get('foo'))